requirements.txt
pydantic
notebook
matplotlib
pandas
anthropic
Pillow
